# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MindForge Personal Second Brain API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tools/summarize:
    post:
      operationId: summarizeDocument
      tags: [tools]
      summary: Generate a summary of a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolDocumentBody"
      responses:
        "200":
          description: Summary result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolResult"
  /tools/action-items:
    post:
      operationId: extractActionItems
      tags: [tools]
      summary: Extract action items from a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolDocumentBody"
      responses:
        "200":
          description: Action items result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolResult"
  /tools/flashcards:
    post:
      operationId: generateFlashcards
      tags: [tools]
      summary: Generate flashcards from a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolDocumentBody"
      responses:
        "200":
          description: Flashcards result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardsResult"
  /tools/mindmap:
    post:
      operationId: generateMindmap
      tags: [tools]
      summary: Generate a Mermaid mind map from a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolDocumentBody"
      responses:
        "200":
          description: Mind map result (Mermaid syntax)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolResult"
...
components:
  schemas:
...
    ToolDocumentBody:
      type: object
      properties:
        documentId:
          type: integer
      required:
        - documentId
    ToolResult:
      type: object
      properties:
        result:
          type: string
        documentId:
          type: integer
      required:
        - result
        - documentId
    Flashcard:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
      required:
        - question
        - answer
    FlashcardsResult:
      type: object
      properties:
        flashcards:
          type: array
          items:
            $ref: "#/components/schemas/Flashcard"
        documentId:
          type: integer
      required:
        - flashcards
        - documentId
...